Removed nullable mappers.
