+++ app/src/main/java/com/kasianov/sergei/omaloma/core/extentions/Event.kt
@@ -1,7 +1,5 @@
 package com.kasianov.sergei.omaloma.core.extentions
 
-import androidx.lifecycle.Observer
-
 /**
  * Used as a wrapper for data that is exposed via a LiveData that represents an event.
  */
@@ -28,16 +26,3 @@ open class Event<out T>(private val content: T) {
     fun peekContent(): T = content
 }
 
-/**
- * An [Observer] for [Event]s, simplifying the pattern of checking if the [Event]'s content has
- * already been handled.
- *
- * [onEventUnhandledContent] is *only* called if the [Event]'s contents has not been handled.
- */
-class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
-    override fun onChanged(event: Event<T>?) {
-        event?.getContentIfNotHandled()?.let {
-            onEventUnhandledContent(it)
-        }
-    }
-}
