+++ app/src/main/java/com/kasianov/sergei/omaloma/core/extentions/EventObserver.kt
@@ -1,0 +1,18 @@
+package com.kasianov.sergei.omaloma.core.extentions
+
+import androidx.lifecycle.Observer
+
+/**
+ * An [Observer] for [Event]s, simplifying the pattern of checking if the [Event]'s content has
+ * already been handled.
+ *
+ * [onEventUnhandledContent] is *only* called if the [Event]'s contents has not been handled.
+ */
+class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) :
+    Observer<Event<T>> {
+    override fun onChanged(event: Event<T>?) {
+        event?.getContentIfNotHandled()?.let {
+            onEventUnhandledContent(it)
+        }
+    }
+}
