+++ app/src/main/java/com/kasianov/sergei/omaloma/core/extentions/ListMapper.kt
@@ -1,0 +1,5 @@
+package com.kasianov.sergei.omaloma.core.extentions
+
+// Non-nullable to non-nullable
+interface ListMapper<I, O> :
+    Mapper<List<I>, List<O>>
